+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changelog
 ------------------
 
 - #2120 Support for Multiple component analysis
+- #2315 Apply dynamic analyses specs for new added analyses
+- #2314 Display error for required fields without value in current language
+- #2313 Log error when calculation fails
+- #2310 Added `get_relative_delta` and `get_tzinfo` in datetime API
+- #2311 Properly process and validate field values from sample header on submit
+- #2307 Rely on fields when validating submitted values on sample creation
 - #2305 Add support for dates in ANSI X3.30 and ANSI X3.43.3 formats
 - #2304 Fix dynamic sample specification not applied for new samples
 - #2303 Fix managed permission of analysis workflow for lab roles